requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
polars>=0.20.0
xarray>=2023.1.0
pydantic>=2.0.0
PyYAML>=6.0

# Machine Learning
torch>=2.0.0
pytorch-lightning>=2.0.0
darts>=0.27.0
scikit-learn>=1.3.0

# Geospatial
geopandas>=0.14.0
rasterio>=1.3.0
shapely>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter (optional for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.0.270

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
